[Phase 0] Fix build configuration and test bugs
- Add MSVC-specific compiler flags for C++17 compliance
- Fix DeviceTest linking by adding missing dependencies
- Fix IpAddressValidator: strict IPv4 validation and qint64 for host count
- Fix StatisticsCalculator test expectations for sample variance/stddev
- All 5 unit tests now passing (100%)

+++ src/utils/IpAddressValidator.h
@@ -17,7 +17,7 @@ public:
     static QPair<QString, int> parseCidr(const QString& cidr);
     static quint32 ipToUInt32(const QString& ip);
     static QString uInt32ToIp(quint32 value);
-    static int calculateHostCount(int prefixLength);
+    static qint64 calculateHostCount(int prefixLength);
 
 private:
     IpAddressValidator() = delete;
